scraper/requirements.txt
# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser Automation (for dynamic pages)
playwright>=1.40.0

# AI API
openai>=1.3.0

# Database
supabase>=2.0.0

# Environment
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dateutil>=2.8.0
